Guard remove_user_type migration with column_exists?

+++ db/migrate/20250601055102_remove_user_type_from_users.rb
@@ -1,5 +1,7 @@
 class RemoveUserTypeFromUsers < ActiveRecord::Migration[7.0]
   def change
-    remove_column :users, :user_type, :string
+    if column_exists?(:users, :user_type)
+      remove_column :users, :user_type, :string
+    end
   end
 end
